+++ examples/ContainerConfig.php
@@ -6,11 +6,17 @@ namespace WoohooLabs\Zen\Examples;
 use WoohooLabs\Zen\Config\AbstractContainerConfig;
 use WoohooLabs\Zen\Config\EntryPoint\ClassEntryPoint;
 use WoohooLabs\Zen\Config\EntryPoint\WildcardEntryPoint;
+use WoohooLabs\Zen\Config\Hint\ContextDependentDefinitionHint;
 use WoohooLabs\Zen\Config\Hint\DefinitionHint;
 use WoohooLabs\Zen\Config\Hint\WildcardHint;
 use WoohooLabs\Zen\Examples\Controller\AnimalController;
+use WoohooLabs\Zen\Examples\Controller\Authentication\AuthenticationController;
+use WoohooLabs\Zen\Examples\Controller\PlantController;
+use WoohooLabs\Zen\Examples\Controller\UserController;
 use WoohooLabs\Zen\Examples\Service\AnimalService2;
 use WoohooLabs\Zen\Examples\Service\AnimalServiceInterface;
+use WoohooLabs\Zen\Examples\Service\AuthenticationService;
+use WoohooLabs\Zen\Examples\Service\ExtendedAuthenticationService;
 use WoohooLabs\Zen\Examples\Service\PlantService;
 use WoohooLabs\Zen\Examples\Service\PlantServiceInterface;
 
@@ -20,12 +26,12 @@ class ContainerConfig extends AbstractContainerConfig
     {
         return [
             ClassEntryPoint::create(AnimalServiceInterface::class)
-                ->autoload()
+                ->disableAutoload()
                 ->fileBased(),
             ClassEntryPoint::create(AnimalController::class)
                 ->autoload()
                 ->fileBased(),
-            new WildcardEntryPoint(__DIR__ . "/Controller"),
+            WildcardEntryPoint::create(__DIR__ . "/Controller"),
         ];
     }
 
@@ -36,7 +42,27 @@ class ContainerConfig extends AbstractContainerConfig
             AnimalServiceInterface::class => AnimalService2::class,
             PlantServiceInterface::class => DefinitionHint::singleton(PlantService::class)
                 ->setParameter("plantType", "sunflower")
-                ->setProperty("plantType", "sunflower")
+                ->setProperty("plantType", "sunflower"),
+            AuthenticationService::class => ContextDependentDefinitionHint::create()
+                ->setClassContext(
+                    DefinitionHint::singleton(AuthenticationService::class),
+                    [
+                        AnimalController::class,
+                    ]
+                )
+                ->setClassContext(
+                    DefinitionHint::prototype(AuthenticationService::class),
+                    [
+                        PlantController::class,
+                    ]
+                )
+                ->setClassContext(
+                    DefinitionHint::singleton(ExtendedAuthenticationService::class),
+                    [
+                        UserController::class,
+                        AuthenticationController::class,
+                    ]
+                )
         ];
     }