+++ src/Config/EntryPoint/WildcardEntryPoint.php
@@ -6,7 +6,7 @@ namespace WoohooLabs\Zen\Config\EntryPoint;
 use WoohooLabs\Zen\Utils\FileSystemUtil;
 use function rtrim;
 
-class WildcardEntryPoint implements EntryPointInterface
+class WildcardEntryPoint extends AbstractEntryPoint
 {
     /**
      * @var string
@@ -18,16 +18,6 @@ class WildcardEntryPoint implements EntryPointInterface
      */
     private $onlyConcreteClasses;
 
-    /**
-     * @var bool
-     */
-    private $autoloaded;
-
-    /**
-     * @var bool
-     */
-    private $fileBased;
-
     public static function create(string $className, bool $onlyConcreteClasses = true): WildcardEntryPoint
     {
         return new WildcardEntryPoint($className, $onlyConcreteClasses);
@@ -37,38 +27,6 @@ class WildcardEntryPoint implements EntryPointInterface
     {
         $this->directoryName = rtrim($directoryName, "\\/");
         $this->onlyConcreteClasses = $onlyConcreteClasses;
-        $this->autoloaded = false;
-        $this->fileBased = false;
-    }
-
-    public function autoload(): WildcardEntryPoint
-    {
-        $this->autoloaded = true;
-
-        return $this;
-    }
-
-    /**
-     * @internal
-     */
-    public function isAutoloaded(): bool
-    {
-        return $this->autoloaded;
-    }
-
-    public function fileBased(): WildcardEntryPoint
-    {
-        $this->fileBased = true;
-
-        return $this;
-    }
-
-    /**
-     * @internal
-     */
-    public function isFileBased(): bool
-    {
-        return $this->fileBased;
     }
 
     /**