+++ tests/Config/EntryPoint/ClassEntryPointTest.php
@@ -12,67 +12,6 @@ class ClassEntryPointTest extends TestCase
     /**
      * @test
      */
-    public function notAutoloadedByDefault()
-    {
-        $entryPoint = ClassEntryPoint::create(EntryPointA::class);
-
-        $isAutoloaded = $entryPoint->isAutoloaded();
-
-        $this->assertFalse($isAutoloaded);
-    }
-
-    /**
-     * @test
-     */
-    public function autoload()
-    {
-        $entryPoint = ClassEntryPoint::create(EntryPointA::class);
-
-        $entryPoint->autoload();
-
-        $this->assertTrue($entryPoint->isAutoloaded());
-    }
-
-    /**
-     * @test
-     */
-    public function notAutoloaded()
-    {
-        $entryPoint = ClassEntryPoint::create(EntryPointA::class);
-
-        $entryPoint->autoload();
-        $entryPoint->notAutoloaded();
-
-        $this->assertFalse($entryPoint->isAutoloaded());
-    }
-
-    /**
-     * @test
-     */
-    public function notFileBasedByDefault()
-    {
-        $entryPoint = ClassEntryPoint::create(EntryPointA::class);
-
-        $isFileBased = $entryPoint->isFileBased();
-
-        $this->assertFalse($isFileBased);
-    }
-
-    /**
-     * @test
-     */
-    public function fileBased()
-    {
-        $entryPoint = ClassEntryPoint::create(EntryPointA::class);
-
-        $entryPoint->fileBased();
-
-        $this->assertTrue($entryPoint->isFileBased());
-    }
-
-    /**
-     * @test
-     */
     public function getClassNames()
     {
         $entryPoint = new ClassEntryPoint(EntryPointA::class);